refactor: move exceptions to common package
