+++ usefull_stuff/messanger/client.py
@@ -8,7 +8,7 @@ import time
 from _socket import SocketType
 
 from common import const
-from exceptions import IncompleteConfigError
+from common.exceptions import IncompleteConfigError
 from common.abstract_socket import AbstractSocket
