+++ usefull_stuff/messanger/common/exceptions.py
